Move location handler to restapi package
