+++ server/http/restapi/restapi.go
@@ -12,6 +12,7 @@ import (
 	"github.com/mattermost/mattermost-plugin-apps/server/apps"
 	"github.com/mattermost/mattermost-plugin-apps/server/constants"
 	"github.com/mattermost/mattermost-plugin-apps/server/store"
+	"github.com/mattermost/mattermost-plugin-apps/server/utils/httputils"
 )
 
 const (
@@ -35,7 +36,20 @@ func Init(router *mux.Router, apps *apps.Service) {
 	}
 
 	subrouter := router.PathPrefix(constants.APIPath).Subrouter()
+	subrouter.HandleFunc("/locations", checkAuthorized(a.handleLocations)).Methods("GET")
 	subrouter.HandleFunc(SubscribePath, a.handleSubscribe).Methods("POST", "DELETE")
+}
+
+func checkAuthorized(f func(http.ResponseWriter, *http.Request, string)) func(http.ResponseWriter, *http.Request) {
+	return func(w http.ResponseWriter, req *http.Request) {
+		actingUserID := req.Header.Get("Mattermost-User-Id")
+		if actingUserID == "" {
+			httputils.WriteUnauthorizedError(w, errors.New("not authorized"))
+			return
+		}
+
+		f(w, req, actingUserID)
+	}
 }
 
 func (a *api) handleSubscribe(w http.ResponseWriter, r *http.Request) {
@@ -82,4 +96,31 @@ func (a *api) handleSubscribe(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
+}
+
+func (a *api) handleLocations(w http.ResponseWriter, req *http.Request, actingUserID string) {
+	userID := req.URL.Query().Get("user_id")
+	if userID == "" {
+		httputils.WriteBadRequestError(w, errors.New("no user id"))
+		return
+	}
+
+	if userID != actingUserID {
+		httputils.WriteUnauthorizedError(w, errors.New("user id is not the same"))
+		return
+	}
+
+	channelID := req.URL.Query().Get("channel_id")
+	if channelID == "" {
+		httputils.WriteBadRequestError(w, errors.New("no channel id"))
+		return
+	}
+
+	locations, err := a.apps.API.GetLocations(userID, channelID)
+	if err != nil {
+		httputils.WriteInternalServerError(w, err)
+		return
+	}
+
+	httputils.WriteJSON(w, locations)
 }
